+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/FileHelper.java
@@ -34,7 +34,7 @@ public class FileHelper {
         this.minioAdapter = minioAdapter;
     }
 
-    String uploadIidmNetworkToMinio(String taskId, Network network, Instant utcInstant) throws IOException {
+    public String uploadIidmNetworkToMinio(String taskId, Network network, Instant utcInstant) throws IOException {
         Path iidmTmpPath = new File(System.getProperty("java.io.tmpdir"), "network.xiidm").toPath();
         network.write("XIIDM", null, iidmTmpPath);
         String iidmNetworkDestinationPath = String.format("%s/networks/%s", taskId, HOURLY_NAME_FORMATTER.format(utcInstant).concat(".xiidm"));
@@ -44,7 +44,7 @@ public class FileHelper {
         return minioAdapter.generatePreSignedUrl(iidmNetworkDestinationPath);
     }
 
-    String uploadJsonCrac(String taskId, Crac crac, Instant utcInstant) {
+    public String uploadJsonCrac(String taskId, Crac crac, Instant utcInstant) {
         String jsonCracFilePath = String.format("%s/cracs/%s", taskId, HOURLY_NAME_FORMATTER.format(utcInstant).concat(".json"));
         ByteArrayOutputStream cracByteArrayOutputStream = new ByteArrayOutputStream();
         CracExporters.exportCrac(crac, "Json", cracByteArrayOutputStream);
@@ -73,7 +73,7 @@ public class FileHelper {
         return cracCreationContext.getCrac();
     }
 
-    String uploadRaoParameters(String taskId, Instant utcInstant) {
+    public String uploadRaoParameters(String taskId, Instant utcInstant) {
         String raoParametersFilePath = String.format("%s/rao-parameters/%s", taskId, HOURLY_NAME_FORMATTER.format(utcInstant).concat(".json"));
         RaoParameters raoParameters = RaoParameters.load();
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
