+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/dichotomy/SweCsaDichotomyRunner.java
@@ -14,16 +14,74 @@ package com.farao_community.farao.swe_csa.app.dichotomy;
 import com.farao_community.farao.commons.EICode;
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_api.range_action.CounterTradeRangeAction;
+import com.farao_community.farao.rao_runner.api.resource.RaoRequest;
 import com.farao_community.farao.rao_runner.api.resource.RaoResponse;
+import com.farao_community.farao.rao_runner.starter.RaoRunnerClient;
+import com.farao_community.farao.swe_csa.api.resource.CsaRequest;
+import com.farao_community.farao.swe_csa.api.resource.CsaResponse;
+import com.farao_community.farao.swe_csa.api.resource.Status;
+import com.farao_community.farao.swe_csa.app.FileHelper;
+import com.farao_community.farao.swe_csa.app.SweCsaRunner;
+import com.farao_community.farao.swe_csa.app.Threadable;
+import com.farao_community.farao.swe_csa.app.ZipHelper;
+import com.farao_community.farao.swe_csa.app.dichotomy.dispatcher.SweCsaShiftDispatcher;
+import com.farao_community.farao.swe_csa.app.dichotomy.index.Index;
+import com.farao_community.farao.swe_csa.app.dichotomy.shifter.LinearScaler;
+import com.farao_community.farao.swe_csa.app.dichotomy.variable.MultipleDichotomyVariables;
 import com.powsybl.iidm.network.Country;
 import com.powsybl.iidm.network.Network;
 import org.apache.commons.lang3.tuple.Pair;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.FileAttribute;
+import java.nio.file.attribute.PosixFilePermission;
+import java.nio.file.attribute.PosixFilePermissions;
+import java.time.Instant;
 import java.util.Map;
+import java.util.Set;
 
-public class DichotomyRunner {
+@Service
+public class SweCsaDichotomyRunner {
 
-    public RaoResponse launchDichotomy(Network network, Crac crac, String timestamp, SweCsaRaoValidator validator) {
+    private final RaoRunnerClient raoRunnerClient;
+    private final FileHelper fileHelper;
+    private static final Logger LOGGER = LoggerFactory.getLogger(SweCsaRunner.class);
+
+    public SweCsaDichotomyRunner(RaoRunnerClient raoRunnerClient, FileHelper fileHelper) {
+        this.raoRunnerClient = raoRunnerClient;
+        this.fileHelper = fileHelper;
+    }
+
+    @Threadable
+    public CsaResponse runRaoDichotomy(CsaRequest csaRequest) throws IOException {
+        String requestId = csaRequest.getId();
+        LOGGER.info("Csa request received : {}", csaRequest);
+        FileAttribute<Set<PosixFilePermission>> attr = PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString("rwx------"));
+        Path archiveTempPath = Files.createTempFile("csa-temp-inputs", "inputs.zip", attr);
+
+        String timestamp = csaRequest.getBusinessTimestamp();
+        Instant utcInstant = Instant.parse(csaRequest.getBusinessTimestamp());
+        ZipHelper.zipDataCsaRequestFiles(csaRequest, archiveTempPath);
+        Network network = fileHelper.importNetwork(archiveTempPath);
+        Crac crac = fileHelper.importCrac(archiveTempPath, network, utcInstant);
+        String networkFileUrl = fileHelper.uploadIidmNetworkToMinio(requestId, network, utcInstant);
+        String cracFileUrl = fileHelper.uploadJsonCrac(requestId, crac, utcInstant);
+        String raoParametersUrl = fileHelper.uploadRaoParameters(requestId, utcInstant);
+        RaoRequest raoRequest = new RaoRequest(requestId, networkFileUrl, cracFileUrl, raoParametersUrl);
+        SweCsaRaoValidator validator = new SweCsaRaoValidator(raoRunnerClient, requestId, networkFileUrl, cracFileUrl, crac, raoParametersUrl);
+
+        RaoResponse raoResultAfterDichotomy = launchDichotomy(network, crac, timestamp, validator);
+        LOGGER.info("dichotomy RAO computation answer received for TimeStamp: '{}'", raoRequest.getInstant());
+
+        return new CsaResponse(raoResultAfterDichotomy.getId(), Status.FINISHED.toString());
+    }
+
+    private RaoResponse launchDichotomy(Network network, Crac crac, String timestamp, SweCsaRaoValidator validator) {
 
         Pair<MultipleDichotomyVariables, MultipleDichotomyVariables> initialDichotomyVariable = getInitialDichotomyIndex(crac);
         DichotomyEngine<RaoResponse, MultipleDichotomyVariables> engine = new DichotomyEngine<>(
