+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/dichotomy/dispatcher/ShiftDispatcher.java
@@ -1,6 +1,7 @@
-package com.farao_community.farao.swe_csa.app.dichotomy;
+package com.farao_community.farao.swe_csa.app.dichotomy.dispatcher;
 
 import com.farao_community.farao.dichotomy.api.exceptions.ShiftingException;
+import com.farao_community.farao.swe_csa.app.dichotomy.variable.DichotomyVariable;
 
 import java.util.Map;
