+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/dichotomy/index/IndexStrategy.java
@@ -1,4 +1,4 @@
-package com.farao_community.farao.swe_csa.app.dichotomy;
+package com.farao_community.farao.swe_csa.app.dichotomy.index;
 
 /*
  * Copyright (c) 2024, RTE (http://www.rte-france.com)
@@ -6,6 +6,8 @@ package com.farao_community.farao.swe_csa.app.dichotomy;
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
+import com.farao_community.farao.swe_csa.app.dichotomy.variable.DichotomyVariable;
 
 /**
  * @author Jean-Pierre Arnould {@literal <jean-pierre.arnould at rte-france.com>}
