+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/dichotomy/shifter/NetworkShifter.java
@@ -1,4 +1,4 @@
-package com.farao_community.farao.swe_csa.app.dichotomy;
+package com.farao_community.farao.swe_csa.app.dichotomy.shifter;
 
 /*
  * Copyright (c) 2024, RTE (http://www.rte-france.com)
@@ -13,6 +13,7 @@ package com.farao_community.farao.swe_csa.app.dichotomy;
 
 import com.farao_community.farao.dichotomy.api.exceptions.GlskLimitationException;
 import com.farao_community.farao.dichotomy.api.exceptions.ShiftingException;
+import com.farao_community.farao.swe_csa.app.dichotomy.variable.DichotomyVariable;
 import com.powsybl.iidm.network.Network;
 
 public interface NetworkShifter<U extends DichotomyVariable<U>> {
